tests/unit/test_performance_metrics.py
"""
Unit tests for SSB Retire Server Web Application performance metrics functionality.
"""

import pytest
import json
import os
from unittest.mock import patch, MagicMock, mock_open


class TestPerformanceMetrics:
    """Test performance metrics collection functionality."""
    
    def test_performance_metrics_import(self):
        """Test performance metrics module can be imported."""
        try:
            from performance.metrics import PerformanceMetricsCollector
            assert PerformanceMetricsCollector is not None
        except ImportError:
            # Module may not exist yet - this is expected
            pytest.skip("Performance metrics module not implemented yet")
    
    @pytest.mark.skipif(not os.path.exists('performance/metrics.py'), 
                       reason="Performance metrics module not yet implemented")
    def test_performance_metrics_collector_initialization(self):
        """Test PerformanceMetricsCollector initialization."""
        from performance.metrics import PerformanceMetricsCollector
        
        collector = PerformanceMetricsCollector()
        assert collector is not None
    
    @pytest.mark.skipif(not os.path.exists('performance/metrics.py'), 
                       reason="Performance metrics module not yet implemented")
    def test_performance_data_collection(self):
        """Test performance data collection."""
        from performance.metrics import PerformanceMetricsCollector
        
        collector = PerformanceMetricsCollector()
        
        # Mock performance data
        sample_data = {
            'execution_time': 120.5,
            'api_calls': 8,
            'hosts_processed': 3,
            'success_rate': 100.0
        }
        
        # This would test the actual collection method
        # Implementation depends on the actual metrics module
        assert True  # Placeholder test
    
    def test_performance_endpoint_exists(self, authenticated_client):
        """Test performance monitoring endpoint exists."""
        response = authenticated_client.get('/api/performance/metrics')
        
        # Endpoint may not exist yet
        assert response.status_code in [200, 404]
        
        if response.status_code == 200:
            data = json.loads(response.data)
            assert isinstance(data, dict)
    
    def test_performance_data_structure(self, mock_performance_metrics):
        """Test performance data structure validation."""
        required_fields = [
            'execution_time',
            'api_calls', 
            'hosts_processed',
            'success_rate',
            'timestamp'
        ]
        
        for field in required_fields:
            assert field in mock_performance_metrics
        
        # Validate data types
        assert isinstance(mock_performance_metrics['execution_time'], (int, float))
        assert isinstance(mock_performance_metrics['api_calls'], int)
        assert isinstance(mock_performance_metrics['hosts_processed'], int)
        assert isinstance(mock_performance_metrics['success_rate'], (int, float))
        assert isinstance(mock_performance_metrics['timestamp'], str)


class TestPerformanceMonitoring:
    """Test performance monitoring functionality."""
    
    def test_response_time_measurement(self, authenticated_client):
        """Test response time measurement for API endpoints."""
        import time
        
        start_time = time.time()
        response = authenticated_client.get('/health')
        end_time = time.time()
        
        response_time = end_time - start_time
        
        assert response.status_code == 200
        assert response_time < 5.0  # Should respond within 5 seconds
        
        # Ideally, should be much faster
        assert response_time < 1.0  # Should respond within 1 second
    
    def test_concurrent_request_handling(self, authenticated_client):
        """Test application can handle concurrent requests."""
        import threading
        import time
        
        results = []
        
        def make_request():
            start_time = time.time()
            response = authenticated_client.get('/health')
            end_time = time.time()
            results.append({
                'status_code': response.status_code,
                'response_time': end_time - start_time
            })
        
        # Create multiple threads for concurrent requests
        threads = []
        for i in range(5):
            thread = threading.Thread(target=make_request)
            threads.append(thread)
        
        # Start all threads
        for thread in threads:
            thread.start()
        
        # Wait for all threads to complete
        for thread in threads:
            thread.join()
        
        # Validate results
        assert len(results) == 5
        for result in results:
            assert result['status_code'] == 200
            assert result['response_time'] < 2.0  # Should handle concurrent requests efficiently
    
    def test_memory_usage_monitoring(self, authenticated_client):
        """Test memory usage doesn't grow excessively."""
        import psutil
        import os
        
        # Get initial memory usage
        process = psutil.Process(os.getpid())
        initial_memory = process.memory_info().rss
        
        # Make multiple requests
        for i in range(10):
            response = authenticated_client.get('/health')
            assert response.status_code == 200
        
        # Get final memory usage
        final_memory = process.memory_info().rss
        memory_increase = final_memory - initial_memory
        
        # Memory increase should be reasonable (less than 50MB)
        assert memory_increase < 50 * 1024 * 1024
    
    def test_aap_api_call_efficiency(self, authenticated_client, mock_aap_api, sample_retirement_data):
        """Test AAP API call efficiency."""
        
        # Count API calls made during job launch
        response = authenticated_client.post('/api/launch-job',
                                           data=json.dumps(sample_retirement_data),
                                           content_type='application/json')
        
        assert response.status_code == 200
        
        # Verify minimal API calls were made
        mock_post = mock_aap_api['mock_post']
        mock_get = mock_aap_api['mock_get']
        
        total_calls = mock_post.call_count + mock_get.call_count
        
        # Should make minimal API calls (ideally 1-2 calls per job launch)
        assert total_calls <= 5


class TestPerformanceRegression:
    """Test performance regression detection."""
    
    def test_baseline_performance_tracking(self, authenticated_client):
        """Test baseline performance tracking."""
        # This would compare against stored baseline metrics
        
        baseline_metrics = {
            'health_endpoint_response_time': 0.1,  # 100ms
            'api_test_connection_response_time': 0.5,  # 500ms
            'api_launch_job_response_time': 2.0  # 2 seconds
        }
        
        # Test health endpoint performance
        import time
        start_time = time.time()
        response = authenticated_client.get('/health')
        health_response_time = time.time() - start_time
        
        assert response.status_code == 200
        
        # Should be within acceptable range of baseline (allowing 50% variance)
        max_acceptable = baseline_metrics['health_endpoint_response_time'] * 1.5
        assert health_response_time <= max_acceptable
    
    def test_performance_threshold_validation(self, mock_performance_metrics):
        """Test performance threshold validation."""
        
        # Define performance thresholds
        thresholds = {
            'execution_time': 300,  # 5 minutes max
            'api_calls': 15,  # Max API calls per operation
            'success_rate': 95.0  # Minimum success rate
        }
        
        # Validate against thresholds
        assert mock_performance_metrics['execution_time'] <= thresholds['execution_time']
        assert mock_performance_metrics['api_calls'] <= thresholds['api_calls']
        assert mock_performance_metrics['success_rate'] >= thresholds['success_rate']
    
    def test_performance_degradation_detection(self):
        """Test performance degradation detection logic."""
        
        # Simulate historical performance data
        historical_data = [
            {'execution_time': 100, 'timestamp': '2025-07-01'},
            {'execution_time': 105, 'timestamp': '2025-07-02'},
            {'execution_time': 110, 'timestamp': '2025-07-03'},
            {'execution_time': 115, 'timestamp': '2025-07-04'}
        ]
        
        # Simulate current performance
        current_performance = {'execution_time': 150, 'timestamp': '2025-07-05'}
        
        # Calculate performance degradation
        baseline = sum(d['execution_time'] for d in historical_data) / len(historical_data)
        degradation_percentage = ((current_performance['execution_time'] - baseline) / baseline) * 100
        
        # Performance degradation should be detected if > 20%
        if degradation_percentage > 20:
            # This would trigger an alert in real implementation
            assert degradation_percentage > 20
        else:
            assert degradation_percentage <= 20


class TestPerformanceOptimization:
    """Test performance optimization validation."""
    
    def test_api_consolidation_effectiveness(self, mock_performance_metrics):
        """Test API consolidation effectiveness."""
        
        # Based on the performance optimization PRP
        # NetBox API calls should be reduced from 7 to 3
        # vCenter API calls should be reduced from 8+ to 2
        
        expected_max_api_calls = 8  # Conservative estimate for optimized system
        
        assert mock_performance_metrics['api_calls'] <= expected_max_api_calls
    
    def test_execution_time_optimization(self, mock_performance_metrics):
        """Test execution time optimization."""
        
        # Based on the 60-70% performance improvement target
        # Single host retirement should complete in 1-2 minutes (optimized)
        # vs 3-5 minutes (original)
        
        max_optimized_time = 150  # 2.5 minutes for single host
        
        if mock_performance_metrics['hosts_processed'] == 1:
            assert mock_performance_metrics['execution_time'] <= max_optimized_time
    
    def test_batch_processing_efficiency(self):
        """Test batch processing efficiency."""
        
        # Simulate batch processing metrics
        batch_metrics = {
            'hosts_processed': 5,
            'execution_time': 600,  # 10 minutes for 5 hosts
            'api_calls': 25  # 5 calls per host average
        }
        
        # Calculate efficiency metrics
        time_per_host = batch_metrics['execution_time'] / batch_metrics['hosts_processed']
        api_calls_per_host = batch_metrics['api_calls'] / batch_metrics['hosts_processed']
        
        # Should show efficiency gains from batch processing
        assert time_per_host <= 150  # 2.5 minutes per host max
        assert api_calls_per_host <= 8  # 8 API calls per host max
    
    def test_parallel_processing_limits(self):
        """Test parallel processing safety limits."""
        
        # Validate that parallel processing respects safety limits
        # Based on enterprise constraints: 5 concurrent host limit
        
        max_concurrent_hosts = 5
        
        # This would test the actual parallel processing implementation
        # For now, validate the concept
        assert max_concurrent_hosts == 5
        
        # In real implementation, would test:
        # - No more than 5 hosts processed simultaneously
        # - Proper resource management
        # - Error handling in parallel operations