requirements-test.txt
# Testing requirements for Flask application
pytest>=7.0.0
pytest-flask>=1.2.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
responses>=0.23.0
factory-boy>=3.2.0
coverage>=7.0.0
flake8>=6.0.0
bandit>=1.7.0
safety>=2.3.0